product -> whiskey refactor
