+++ next-app/src/app/(app-layout)/chemicals/chemicals/page.tsx
@@ -1,14 +1,8 @@
-import { Container } from '@/common/container';
-import { APP_DESCRIPTION, APP_TITLE } from '@/common/common-utils';
-import { routes } from '@/routing/routing-utils';
-import { ButtonLink } from '@/common/button-link';
-import { getMetadata } from '@/seo/seo-utils';
-import Image from 'next/image';
-import type { Whiskey } from '@/products/product-types';
-import { PageTitle } from '@/common/page-title';
 import { Paper } from '@/common/paper';
 import { Section } from '@/common/section';
-import {Chemical} from '@/products/product-types';
+import { getMetadata } from '@/seo/seo-utils';
+import { Chemical } from '@/whiskeys/object-types';
+import Image from 'next/image';
 
 type WhiskeyDetailsProps = {
   chemical: Chemical;
