+++ next-app/src/app/(app-layout)/flavours/[flavourId]/page.tsx
@@ -1,11 +1,11 @@
 import { PageTitle } from '@/common/page-title';
 import { Paper } from '@/common/paper';
 import { Section, SectionTitle } from '@/common/section';
-import { WhiskeyDetails } from '@/products/product-details';
-import { getOneWhiskeyById } from '@/products/product-fetchers';
-import { WhiskeyGridSkeleton } from '@/products/product-grid';
-import { RelatedProducts } from '@/products/related-products';
 import { getMetadata } from '@/seo/seo-utils';
+import { RelatedProducts } from '@/whiskeys/related-whiskey';
+import { WhiskeyDetails } from '@/whiskeys/whiskey-details';
+import { getOneWhiskeyById } from '@/whiskeys/whiskey-fetcher';
+import { WhiskeyGridSkeleton } from '@/whiskeys/whiskey-grid';
 import type { Metadata } from 'next';
 import { notFound } from 'next/navigation';
 import { Suspense } from 'react';
